+++ CHANGES.md
@@ -1,7 +1,14 @@
+## HEKO-1.0.2
+#### ADDED
+- move playlist
+- remove playlist
+- arrow on playlists that contains other playlists
+
 ## HEKO-1.0.1
 #### ADDED
 - line between table head and table body on current playlist song list
 - current song details in footer
+- context menus
 #### FIXED
 - play / pause button when listening a song by clicking one on the current playlist
 - scroll bar in current playlist song list